Add files via upload
Did some calculations with regards to bullet landing. Used a Boolean to see if the user is holding down their mouse button and if so then I will update the attributes of the bullet class and then calculate the landing coordinates of the bullet so that when I apply the sprite production it will work as intended.