Add files via upload
Worked on Projectile motion and utilizing the landing coordinate calculation I had made prior. I need to work on normalizing the rate at which the bullets are shot because it is more of a burst instead of a fluid motion. I also need to optimize because the landing coordinate calculation cannot keep up when the character is moving.