Add files via upload
Made some random deviation of the bullet so that the shots look more organic. Planned out the different attacks the enemy will do. Organized my code and got rid of things I don’t need anymore. Calculated the elapsed time in seconds based on the amount of frames that have gone by. Based on the time elapsed I can make an invincibility period for the player so that they don’t keep taking damage if they collide with the boss sprite.